Changed Authors
As this is a group project who all put in time and effort. James did not do all of the work. It was his idea and he undid all the work that was completed without him, It is still a group project and authors should depict that. I still have James as Primary architect, and David who did the second mostwork as Secondary Architect. And finally Ary and I as Assistants.